requirements.txt
bedrock-agentcore
strands-agents
strands-agents-tools
boto3
requests
mcp
